+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `f-slider` 组件废除 `range` 配置项
 - `f-slider` 组件新增 `background` `active-color` `width` 配置项
 - `f-card` 组件新增 `border-color` `background` `title-color` `title-bold` `close` `on-close` 配置项
+- `f-input` 组件新增 `autocomplete` 配置项
 
 ## 0.22.1 (2023-02-07)